feat: Add redirect to main page
Add a new route to the Flask application that redirects the root URL to the "apidocs/" endpoint. This will provide a more user-friendly experience when accessing the application.

The redirect is implemented using the `redirect` function from the Flask library, and the HTTP status code `HTTPStatus.OK` is returned to indicate a successful redirection.

This commit addresses the need to improve the user experience by providing a clear entry point to the application.

Refs: #N/A