Refactor authentication module and update dependencies
This commit refactors the authentication module by replacing the `SearchQuery` class with the `BasicAuth` class. The `SearchQuery` class, which represented the search query, has been removed. Instead, the `BasicAuth` class is introduced to handle basic authentication. The `parse_auth_string` method is added to parse the authentication string and extract the username and password. Additionally, the requirements.txt file is updated to use the `quart` and `quart-cors` packages instead of `flask[async]` and `flask_httpauth`. The `flasgger` package is also pinned to version 0.9.7.1. Finally, the `fertiscan-pipeline` package is added as a dependency from the GitHub repository.